[SVLS-5262] redistribute span pointer functions
